SW-1352 Refactor version_comparator.py and COMPARISON_OPTIONS into Util package
